requirements.txt
streamlit>=1.33.0
pandas>=2.0.0
plotly>=5.18.0
openai>=1.23.0
pdfkit>=1.0.0
python-dotenv>=1.0.1
appwrite>=11.0.0
pyarrow==12.0.1
kaleido>=0.2.1
